small improvements and cleanup for Commit()
