+++ ocflv1/store_test.go
@@ -8,12 +8,9 @@ import (
 	"path/filepath"
 	"strings"
 	"testing"
-	"testing/fstest"
 
 	"github.com/srerickson/ocfl"
-	"github.com/srerickson/ocfl/digest"
 	"github.com/srerickson/ocfl/extensions"
-	"github.com/srerickson/ocfl/internal/testfs"
 	"github.com/srerickson/ocfl/ocflv1"
 )
 
@@ -23,6 +20,23 @@ type storeTest struct {
 	name   string
 	size   int
 	layout extensions.Layout
+}
+
+// layout used for fixture stores w/o layout file
+var testStoreLayout customLayout
+
+type customLayout struct{}
+
+var _ extensions.Layout = (*customLayout)(nil)
+
+func (l customLayout) Name() string {
+	return "uri-encode"
+}
+
+func (l customLayout) NewFunc() (extensions.LayoutFunc, error) {
+	return func(id string) (string, error) {
+		return url.QueryEscape(id), nil
+	}, nil
 }
 
 func TestGetStore(t *testing.T) {
@@ -88,215 +102,4 @@ func TestGetStore(t *testing.T) {
 			}
 		})
 	}
-}
-
-func TestScanObjects(t *testing.T) {
-	ctx := context.Background()
-
-	// map to store to expected # of objects
-	var storeTests = []storeTest{
-		{name: `good-stores/reg-extension-dir-root`, size: 1, layout: nil},
-		{name: `good-stores/unreg-extension-dir-root`, size: 1, layout: testStoreLayout},
-		{name: `good-stores/simple-root`, size: 3, layout: testStoreLayout},
-		{name: `warn-stores/fedora-root.zip`, size: 176, layout: testStoreLayout},
-		{name: `bad-stores/E072_root_with_file_not_in_object`, size: 1, layout: testStoreLayout},
-		{name: `bad-stores/E073_root_with_empty_dir.zip`, size: 0, layout: testStoreLayout},
-	}
-	optTable := map[string]*ocflv1.ScanObjectsOpts{
-		`default`:       nil,
-		`validate`:      {Strict: true},
-		`no-validate`:   {Strict: false},
-		`fast`:          {Concurrency: 16},
-		`slow`:          {Concurrency: 1},
-		`fast-validate`: {Strict: true, Concurrency: 16},
-		`slow-validate`: {Strict: true, Concurrency: 1},
-		// `w-timeout`: {Timeout: time.Microsecond}, // not sure how to test this..
-	}
-	for mode, opt := range optTable {
-		t.Run(mode, func(t *testing.T) {
-			for _, sttest := range storeTests {
-				t.Run(sttest.name, func(t *testing.T) {
-					var fsys ocfl.FS
-					var root string
-					var store *ocflv1.Store
-					var err error
-					var expectErr bool
-					if opt != nil && opt.Strict && strings.HasPrefix(sttest.name, "bad-stores") {
-						expectErr = true
-					}
-					if strings.HasSuffix(sttest.name, `.zip`) {
-						root = "."
-						zreader, err := zip.OpenReader(filepath.Join(storePath, sttest.name))
-						if err != nil {
-							t.Fatal(err)
-						}
-						defer zreader.Close()
-						fsys = ocfl.NewFS(zreader)
-					} else {
-						fsys = ocfl.NewFS(os.DirFS(storePath))
-						root = sttest.name
-					}
-					store, err = ocflv1.GetStore(ctx, fsys, root)
-					if err != nil {
-						t.Fatal(err)
-					}
-					numObjs := 0
-					scanFn := func(obj *ocflv1.Object) error {
-						if _, err := obj.Inventory(ctx); err != nil {
-							t.Fatal(err)
-						}
-						numObjs++
-						return nil
-					}
-					scanErr := store.ScanObjects(ctx, scanFn, opt)
-					if expectErr {
-						if scanErr == nil {
-							t.Fatal("expected an error")
-						}
-						return
-					}
-					if scanErr != nil {
-						t.Fatal(scanErr)
-					}
-					if numObjs != sttest.size {
-						t.Fatalf("expected %d objects, got %d", sttest.size, numObjs)
-					}
-				})
-			}
-		})
-	}
-}
-
-func TestStoreUpdateObject(t *testing.T) {
-	storePath := "test-stage"
-	ctx := context.Background()
-	storeFS := testfs.NewMemFS() // store
-	stgFS := stageFS()
-	// initialize store
-	if err := ocflv1.InitStore(ctx, storeFS, storePath, nil); err != nil {
-		t.Fatal(err)
-	}
-	store, err := ocflv1.GetStore(ctx, storeFS, storePath)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// v1
-	stage := ocfl.NewStage(digest.SHA256(), ocfl.StageRoot(stgFS, `src1`))
-	if err := stage.AddAllFromRoot(ctx); err != nil {
-		t.Fatal(err)
-	}
-	if _, _, err := stage.GetVal("tmp.txt"); err != nil {
-		t.Fatal(err)
-	}
-	if err = store.Commit(ctx, "object-1", stage,
-		ocflv1.WithContentDir("foo"),
-		ocflv1.WithVersionPadding(2),
-		ocflv1.WithUser("Bill", "mailto:me@no.com"),
-		ocflv1.WithMessage("first commit"),
-	); err != nil {
-		t.Fatal(err)
-	}
-	obj, err := store.GetObject(ctx, "object-1")
-	if err != nil {
-		t.Fatal(err)
-	}
-	if result := obj.Validate(ctx); result.Err() != nil {
-		t.Fatal("object is invalid", result.Err())
-	}
-	inv, err := obj.Inventory(ctx)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if inv.ContentDirectory != "foo" {
-		t.Fatal("expected foo")
-	}
-	if inv.DigestAlgorithm != digest.SHA256id {
-		t.Fatalf("expected sha256")
-	}
-	if inv.Head.Padding() != 2 {
-		t.Fatalf("expected 2")
-	}
-	if u := inv.Versions[inv.Head].User; u == nil || u.Name != "Bill" {
-		t.Fatal("expected Bill")
-	}
-	if u := inv.Versions[inv.Head].User; u == nil || u.Address != "mailto:me@no.com" {
-		t.Fatal("expected Bill")
-	}
-	if inv.Versions[inv.Head].Message != "first commit" {
-		t.Fatal("expected 'first commit'")
-	}
-	// stage 2
-	stage2, err := obj.NewStage(ctx, ocfl.Head)
-	if err != nil {
-		t.Fatal()
-	}
-	if err := stage2.Remove("tmp.txt"); err != nil {
-		t.Fatal(err)
-	}
-	if err := store.Commit(ctx, "object-1", stage2); err != nil {
-		t.Fatal(err)
-	}
-	obj, err = store.GetObject(ctx, "object-1")
-	if err != nil {
-		t.Fatal(err)
-	}
-	if result := obj.Validate(ctx); result.Err() != nil {
-		t.Fatal("object is invalid", result.Err())
-	}
-	inv, err = obj.Inventory(ctx)
-	if err != nil {
-		t.Fatal(err)
-	}
-	head, err := inv.Index(ocfl.Head)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if head.Len() != 0 {
-		t.Fatal("expected head state to be empty")
-	}
-
-	// v3
-	stage3, err := obj.NewStage(ctx, ocfl.Head)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if err := store.Commit(ctx, "object-1", stage3); err != nil {
-		t.Fatal(err)
-	}
-	obj, err = store.GetObject(ctx, "object-1")
-	if err != nil {
-		t.Fatal(err)
-	}
-	if result := obj.Validate(ctx); result.Err() != nil {
-		t.Fatal("object is invalid", result.Err())
-	}
-}
-
-// layout used for fixture stores w/o layout file
-var testStoreLayout customLayout
-
-type customLayout struct{}
-
-var _ extensions.Layout = (*customLayout)(nil)
-
-func (l customLayout) Name() string {
-	return "uri-encode"
-}
-
-func (l customLayout) NewFunc() (extensions.LayoutFunc, error) {
-	return func(id string) (string, error) {
-		return url.QueryEscape(id), nil
-	}, nil
-}
-
-func stageFS() ocfl.FS {
-	fsys := fstest.MapFS{
-		`src1/tmp.txt`:       &fstest.MapFile{Data: []byte(`content1`)},
-		`src2/a/tmp.txt`:     &fstest.MapFile{Data: []byte(`content2`)},
-		`src3/a/tmp.txt`:     &fstest.MapFile{Data: []byte(`content3`)},
-		`src3/b/another.txt`: &fstest.MapFile{Data: []byte(`another`)},
-		`src4/b/another.txt`: &fstest.MapFile{Data: []byte(`another`)},
-	}
-	return ocfl.NewFS(fsys)
 }
